Refactor alterFile into its own module
